x64 port
git-svn-id: http://miranda.googlecode.com/svn/trunk@9306 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/db3x/dbmodulechain.c
@@ -83,7 +83,7 @@ static DWORD FindExistingModuleNameOfs(const char *szName,int nameLen)
 DWORD GetModuleNameOfs(const char *szName)
 {
 	struct DBModuleName dbmn;
-	int nameLen=strlen(szName);
+	int nameLen=(int)strlen(szName);
 	DWORD ofsNew,ofsExisting;
 
 	ofsExisting=FindExistingModuleNameOfs(szName,nameLen);
